requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pydantic>=2.6
python-dotenv
requests
pymysql
